+++ sdl3-image-sys/ChangeLog.md
@@ -4,7 +4,7 @@
     - Update sdl3-sys to 0.6.0
     - Synchronize version number with sdl3-sys
     - Add metadata
-    - Make some functions safe
+    - Make some functions safe and/or const
 
 - 0.2.0:
     - Update sdl3-sys to 0.5.0