+++ sdl3-sys/ChangeLog.md
@@ -5,7 +5,7 @@
     - Add metadata
     - Remove deprecated aliases
     - Change return type of `SDL_Vulkan_GetVkGetInstanceProcAddr`
-    - Make some functions safe
+    - Make some functions safe and/or const
     - MSRV 1.85
 
 - 0.5.4: